playing around with changes to values.hpp

+++ src/Values.hpp
@@ -5,8 +5,10 @@
 
 namespace values {
 
-    const int MESSAGE_SIZE = 1024;
+    // length of user-generated message (including DEAD_SPACE)
+    const size_t MESSAGE_SIZE = 1024;
 
+    // length of user-generated message plus enough space for crc (32 bit int saved in 4 chars)
     const int MESSAGE_WITH_CRC_SIZE = MESSAGE_SIZE + 4;
 
     // msg_type + ip + device_name + "data" + crc
